+++ src/opt_cbody.c++
@@ -9,7 +9,7 @@ CBody::CBody(double Mass, double Radius, double Speed) {
 	radius = Radius;
 	v.x = Speed;
 }
-CBody::CBody(double Mass, double Radius, double Speed, double X, double Y, double Z, int H) {
+CBody::CBody(double Mass, double Radius, double Speed, double X, double Y, double Z, double H) {
 	init();
 	h = H;
 	r = Pos(X, Y, Z);
